+++ 자이-백병재/2주차/todo_list/src/components/Todo_List.tsx
@@ -1,25 +1,19 @@
 import React, { useState, type JSX } from "react";
 import type { Items } from "../types/Items";
 import List_Component from "./List_component";
+import Input_Form from "./Input_Form";
 
 const Todo_List = (): JSX.Element => {
     const [todo_list, setTodo_list] = useState<Items[]>([]);    // useState를 이용, Items[]형식의 todo_list
     const [done_list, setDone_list] = useState<Items[]>([]);    // useState를 이용, Items[]형식의 todo_list
-    const [input, setInput] = useState("");                     // 입력용 state
 
-    const getInput = (e: React.FormEvent) => {                  // 입력받기
-        e.preventDefault();                                     // 새로고침 방지
-        if(!input.trim()) return;                               // trim을 사용, 값이 없으면 리턴
-
+    const addTodoList = (text: string) => {                  // 입력받기
         const newTodo: Items = {                                // 입력값을 newTodo로 가져옴
             id: Date.now(),
-            text: input,
+            text,
             complete: false,
         };
-
         setTodo_list([...todo_list, newTodo]);                  // todo_list에 newTodo값 추가
-        setInput("");                                           // input을 ""로 초기화
-
     }
 
     const doneTodo = (id: number) => {                                  // 완료하기
@@ -41,11 +35,7 @@ const Todo_List = (): JSX.Element => {
     return (
     <div id="root_container">
         <h1 id="root_title">Todo-List</h1>
-        <form id="root_form" onSubmit={getInput}>
-            <input id="input_text" type="text" placeholder="할 일을 입력하세요"
-            value={input} onChange={(e) => setInput(e.target.value)} />
-            <button id="input_button" type="submit">추가</button>
-        </form>
+        <Input_Form addTodo={addTodoList} />
         <hr />
         <div id="list_wrapper">
             {/* List_Component 사용해 컴포넌트 분리 */}
